Rename task that sets the "smr" mysql user password
When the mysql container is started for the first time, it takes
care of setting the "smr" user's password based on the MYSQL_PASSWORD
variable in the `.env` file. However, since we change change the
password on each redeploy (for security reasons), when the database
is already initialized with the old password, we need to alter the
database to reflect the password change.