Quote numbers in cron tasks
As of Ansible 2.8, a warning is emitted if the type given in the
playbook is implicitly converted to the type expected by the task.

Unfortunately, since the cron task's timing options (hour, minute,
day, etc.) expect strings in order to support values like "*", we
will get a warning when we specify an integer. For example,

cron:
  hour: 2

will emit a warning like

[WARNING]: The value 2 (type int) in a string field was converted
to u'2' (type string). If this does not look like what you expect,
quote the entire value to ensure it does not change.

To suppress this warning, we convert all numerical arguments in cron
tasks to strings by quoting them.