deploy.py: make playbook run verbose
Ansible playbooks can be run with different levels of verbosity.
Adding one level of verbosity provides very helpful information
about what is being changed that cannot be determined retroactively
if verbosity is off.

It can be a little noisy (depending on what each task chooses to
output), but is well worth the clutter when you need the information.