setup-smr: conditionally run s3 get
Restoring files (e.g. database backups, image uploads) from s3 only
needs to be performed on a newly provisioned machine, so we can skip
this task if the files already exist. We will use one of the history
database files as a proxy for all of the expected s3 files, since it
will always have the same filename.

Note that this won't work as currently implemented. We get the
following error:

ERROR: S3 error: 403 (InvalidObjectState): The operation is not valid for the object's storage class

This is because the files we're trying to access have already been
transferred to glacier. We're going to need to skip glacier objects,
but that's not an option for s3cmd -- we'll likely need to use the
AWS CLI directly (once an official Docker image is available, see
https://github.com/aws/aws-cli/issues/3553).