Update SMTP_HOSTNAME in live config.specific.php
Refer to the smtp container name in the dockerize (live) repo. This
name is unique ("smr-smtp"), unlike the service name ("smtp"), which
is the same in both beta and live. This ensures we only ever connect
to the live smtp service.

This fixes a bug where some pages were connecting to the beta smtp
service, and then mx.datenschleuder.org (our mailserver) was rejecting
the messages with the following error because don't set the system
hostname in the smr repo smtp service environment:

Helo command rejected: need fully-qualified hostname