Refactor secret variables to minimize encryption
This allows us to keep configuration files largely unencrypted, which
makes it easier to modify, validate, and track the history of these
files. The encrypted secrets file is in YAML format and only contains
the variables that need to be kept private. These variables are then
substituted into the config files using the `template` module.

It would have been extremely convenient if the `copy` module had a
"template: yes/no" option, because it did an otherwise perfect job of
copying the config files in the simplest way possible.

Instead, we need to use `template` and a `with_filetree` looping
mechanism. This is not ideal, for at least two reasons:

1. `template` can't handle encrypted files (e.g. opendkim), which need
   to be handled by a separate `copy` task.

2. `with_filetree` is too verbose, and needs to be explicitly told not
   to process directories.

We have a pair of variable files: `vault.yml` for the private vars,
and `main.yml` for the public vars. Each private var is prefixed with
`vault_` and then redefined without the prefix as a public var for
easier auditing of the variable list.

This also gets rid of two fragile `setup-smr` tasks, which were being
used to manually replace the MySQL password in configuration files.

NOTE: All variables that were previously encrypted but are now public
have been carefully checked to make sure we're not exposing something
that wasn't already exposed to some public interface.