Fix database restore tasks for MySQL 8.0+
The database restore tasks were run conditionally on the existence of
`game.frm` database files. However, in MySQL 8.0, the `.frm` files
have been removed, which caused the database restore tasks to run on
every deployment. This is fixed by instead looking for the existence
of `game.MYD` files.

Note that we are still relying on MySQL internals here, so this could
break at any time without warning.