Require a .ansible_secret file
Specify the vault password file in ansible.cfg so that we do not need
to pass it explicitly to every ansible-vault and ansible-playbook
command we invoke.

The downside is that, because the path to the password file must be
hard-coded, we can no longer support the `~/.ansible_einstein.secret`
location as well.