setup-webboard: update config.php
Change from an empty table prefix to `phpbb_`. This is needed because
phpBB (as of 3.3.4) has a table named `groups` that conflicts with the
MySQL 8.0 reserved keyword `groups` if no table prefix is used.

I manually modified all tables to add a `phpbb_` prefix. This solved
the problem once config.php was updated and the cache directory was
cleared of all files (except .htaccess and index.htm).

This issue is tracked at https://tracker.phpbb.com/browse/PHPBB3-16830.