einstein.yml: install the latest Docker
We got quite far behind in Docker version (18.09.0), and this was
causing some docker errors for beta (where we build on-site).

Use `docker_package_state: latest` instead of `present` so that we
upgrade to the latest version. Ideally we would pin a specific
version, but it seems a bit tricky because you have to specify the
precise package name (see e.g.
https://github.com/geerlingguy/ansible-role-docker/issues/45).