Make file modes explicit
It is a best security practice to make file permission modes explicit.

Fixes ansible-lint warning:
> risky-file-permissions: File permissions unset or incorrect.

Sometimes we want to use the existing remote file mode (such as when a
later task or some remote process modifies the original mode), and the
only way to do this is to omit `mode` and mark them with "noqa".