Skip no-changed-when lint for docker commands
There is a "community.docker" ansible collection, which if used would
provide proper tracking of task idempotence; however, it seems to have
some uncertainty about its future (especially w.r.t. the API it uses
to connect to Docker, with some 3rd party tools becoming deprecated
and internal changes in Docker due to buildx and porting compose from
python to go in v2.0.0). I want to wait for the dust to settle a bit
before adopting this collection.

Since we are sticking with our simple `docker` and `docker-compose`
commands (at least for now), we have no easy way to check idempotence;
thus just skip the `no-changed-when` lint rule.