beta: switch to https protocol for github
We were using the git protocol to make it easier to push new code
directly from the remote host; however, this relies on either having
a private SSH key on the remote host (not recommended), or having SSH
forwarding set up in ansible.

Note that the git protocol works for einstein only because it is set
up to allow logging in directly to the root user. On a Vagrant VM of
Ubuntu 18.04, you cannot log in directly as root (without sshd mods)
and so when you "become" root, you lose your SSH agent.

It is just much simpler to use the https protocol, and forego push
access via key pair on the remote host.