Add Vagrantfile for local deployment test
We can simulate a full deployment using a local Vagrant VM. Please be
aware that this is currently _identical_ to a production deployment,
and is not safe to blindly run. There are collisions with services run
on 3rd party platforms (Discord and IRC), and the `smr_live` backup
cron can make a potentially destructive backup to the AWS s3 bucket.
The long-term goal is to make this safe to run without compromising
the power of the test.

More work is also needed to inspect the resulting Vagrant VM. We would
like to be able to expose the services in such a way that we can test
the resulting webservers on a browser from the local host. For now, we
can only `vagrant ssh` and check the health of the services.