requirements.yml: add community collections
Third-party collections need to be specified in the requirements.yml
file for them to recognized by ansible-lint (and to correctly enumerate
our dependencies, though `community.general` is likely going to be
present in any ansible installation).

Fixes the following ansible-lint warning:

> WARNING  Unable to load module community.general.timezone at roles/system_config/tasks/main.yml:46 for options validation
> WARNING  Unable to resolve FQCN for module community.general.timezone

We also add the `role` subcommand to the `ansible-galaxy` command in
deploy.py to avoid an informational warning about not installing the
collections in the requirements.yml file.