ansible-lint.yml: don't restrict variable names
In principle it seems like a good idea to prefix variable names with
roles, but our role names are already a bit too cumbersome, so this
makes the variable names even worse. We already have a nice informal
system for variable names, so we'll keep it that way for now.